requirements.txt
anyio==4.0.0
click==8.1.7
exceptiongroup==1.1.3
graphql-core==3.2.3
h11==0.14.0
idna==3.4
libcst==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
Pygments==2.16.1
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0.1
rich==13.5.2
six==1.16.0
sniffio==1.3.0
starlette==0.31.1
strawberry-graphql==0.205.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
uvicorn==0.23.2
